requirements.txt
Django==1.5
South==1.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-autoslug-field==0.2.3
django-crispy-forms==1.4.0
django-gravatar2==1.1.4
django-toolbelt==0.0.1
gunicorn==19.1.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
